vizzys fgd (except icons)
This is every change my updated .fgd makes aside from a few key differences:

This is based on the "mapbase" branch of the mapbase-source/source-fgds repo, which I have changed to be the main branch (not the super outdated main branch.)

I did not add custom Icons/entity model helpers in this specific commit. They will be in the next commit. This is because I'm unsure if you wish to include these.

I did not include the custom soundscapes lines I added. They will be in the next commit. This is because I'm unsure if you wish to include these.

"Touch Opens" flag for the base door class was changed from 1 to 0, not sure is that's intentional on your part but imo its better off by default.

as for the changelog and icons, consider downloading them here:
https://gamebanana.com/tools/14943
(however its slightly outdated compared to this commit)

Whats new?

many missing inputs/outputs
many improved entity descriptions
many entity helper lines/spheres
many typo fixes